// backend/src/types.ts
export interface Employee {
  EmpID: string;
  FirstName: string;
  LastName: string;
  FullName: string;
  Email: string;
  Phone: string;
  Department: string;
  DepartmentCode: string;
  Designation: string;
  Manager: string;
  Location: string;
  EmploymentType: string;
  Status: string;
  DateOfBirth: string;
  DateOfJoining: string;
  AnnualSalary: number;
  PerformanceRating: number;
  LeaveBalance: number;
}